Add files via upload
Developed a responsive webpage dedicated to the Seattle Seahawks using HTML for structure and CSS for styling. Designed the layout to be visually appealing and user-friendly, implementing modern web design principles such as flexbox for alignment and media queries for mobile responsiveness. The project enhanced my skills in front-end development and web design best practices.